Improve clarity of README.md
We were trying to use dotenv at my company, and noticed that we weren't sure from the README how .env/.env.<environment name> files are loaded when using Rails different environments. Therefore, we had to carefully examine which env vars were loaded in different environments using "pp ENV" on the Rails console. This PR improves the README so that others don't have to go through the same process of trying to understand how dotenv behaves in this regard.