Reorder to make the load method available
We are getting the following error when trying to use `require "dotenv/rails-now"`

> /Users/mikaelhenriksson/.rvm/gems/ruby-2.1.8/gems/activesupport-3.2.22.6/lib/active_support/dependencies.rb:243:in `load': wrong number of arguments (0 for 1..2) (ArgumentError)
>	from /Users/mikaelhenriksson/.rvm/gems/ruby-2.1.8/gems/dotenv-rails-2.0.1/lib/dotenv/rails.rb:20:in `block in <class:Railtie>'

Reordering the file so that the methods are defined before using them fixes this problem. Our codebase is from 2005 so I am sure it could be fixed in other ways but this is verified to fix our problem.