don't load .env.local in rails' test environment
In my opinion, the .env.local should not overwrite .env.test
variables. When using .env.local for tests, the tests are not
transparent and can't be executed for example on a CI server.

As the documentation says, the `.env.local` is here for **local**
overwrites. They should not overwrite a test specific variable. Tests
should be run on every machine in every environment and shouldn't
require any specification of environment variables. (My opinion)