Automate Rubygems Releases
This adds a GitHub Actions workflow to Dotenv for code pushes.  If a tag is pushed with a version, both the dotenv and dotenv-rails gems will be published to Rubygems.org.  Note: since Actions only supports 'push' events in the beta, you have to tag and specifically push the tag, instead of using `git push --tags` for this to work.